re-enable cron workflow
tbh there was no reason for me to disable the workflow, but i really (*really*) didn't know if the issue i had was fixable...

again, if the mgs developers are reading this, please create an api for roblox leaderboard. if you implement one, i will discontinue this project as soon as i can.